feat: JpaRepository의 Long -> Integer 변경
- 판단기준이 int 타입의 PlanId이어서 변경

+++ src/main/java/org/backrow/solt/repository/PlanRepository.java
@@ -3,5 +3,5 @@ package org.backrow.solt.repository;
 import org.backrow.solt.domain.Plan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanRepository extends JpaRepository<Plan, Long> {
+public interface PlanRepository extends JpaRepository<Plan, Integer> {
 }
